Snake and Ladder Game (Python)
I have developed this game using Pythons Pygame module, this project was my kind of first moderate level project it took few days to complete this project as I was a beginner as I have just learn the pygame module. In this project required lots of logic but this project helped me a lot to progress in coding.